Split motion_workflow actions into seperate files
