+++ openslides_backend/action/motion_workflow/update.py
@@ -1,0 +1,16 @@
+from ...models.motion_workflow import MotionWorkflow
+from ..action import register_action
+from ..default_schema import DefaultSchema
+from ..generics import UpdateAction
+
+
+@register_action("motion_workflow.update")
+class MotionWorkflowUpdateAction(UpdateAction):
+    """
+    Action to update a motion workflow.
+    """
+
+    model = MotionWorkflow()
+    schema = DefaultSchema(MotionWorkflow()).get_update_schema(
+        properties=["name", "first_state_id"]
+    )
